# gin-backend/docs/swagger.yaml
basePath: /
definitions:
  core.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  core.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/core.ErrorDetail'
    type: object
...
  domain.CreateDomicilioRequest:
    properties:
      alias:
        example: Casa
        type: string
      calle:
        example: Av. Segunda Sur Ote.
        type: string
      colonia_id:
        example: 1
        type: integer
      direccion:
        example: Av. Segunda Sur Ote. 448, Centro, Suchiapa
        type: string
      numero:
        example: "448"
        type: string
      referencia:
        example: Frente al parque
        type: string
      usuario_id:
        example: 100
        type: integer
    type: object
...
  main.CreateAlertaRequest:
    properties:
      mensaje:
        example: El camión llegará 30 min tarde
        type: string
      titulo:
        example: Retraso en ruta
        type: string
      usuario_id:
        example: 100
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para gestión de recolección de residuos
  title: API Recolecta
  version: "1.0"
paths:
  /api/alertas:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar mis alertas
      tags:
      - AlertaUsuario
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload JSON
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.CreateAlertaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear alerta de usuario
      tags:
      - AlertaUsuario
  /api/alertas/{id}/leida:
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Marcar alerta como leída
      tags:
      - AlertaUsuario
...
  /api/domicilios:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar domicilios del usuario autenticado
      tags:
      - Domicilio
    post:
      consumes:
      - application/json
      description: Requiere colonia_id, alias y direccion (o calle). usuario_id es
        opcional si hay JWT.
      parameters:
      - description: Datos del domicilio
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateDomicilioRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear domicilio
      tags:
      - Domicilio
  /api/domicilios/{id}:
    delete:
      parameters:
      - description: ID del domicilio
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar domicilio
      tags:
      - Domicilio
    get:
      parameters:
      - description: ID del domicilio
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener domicilio por ID
      tags:
      - Domicilio
    put:
      consumes:
      - application/json
      parameters:
      - description: ID del domicilio
        in: path
        name: id
        required: true
        type: integer
      - description: Datos del domicilio
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateDomicilioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar domicilio
      tags:
      - Domicilio
...
  /domicilios:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar domicilios (compat)
      tags:
      - Domicilio
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload JSON
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateDomicilioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear domicilio (compat)
      tags:
      - Domicilio
  /domicilios/{id}:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar domicilio (compat)
      tags:
      - Domicilio
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener domicilio (compat)
      tags:
      - Domicilio
    put:
      consumes:
      - application/json
      parameters:
      - description: Payload JSON
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateDomicilioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar domicilio (compat)
      tags:
      - Domicilio
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Ingresa el token JWT con el prefijo Bearer. Ejemplo: Bearer eyJhbGciOiJIUzI1NiIs...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"